refactor(ios): extract ExitNodeSummaryRow into its own file
