+++ NetBird/Source/App/Views/RouteTabView.swift
@@ -81,33 +81,6 @@ struct RouteListView: View {
     }
 }
 
-struct ExitNodeSummaryRow: View {
-    @ObservedObject var routeViewModel: RoutesViewModel
-
-    var body: some View {
-        NavigationLink(destination: ExitNodeSelectionView(routeViewModel: routeViewModel)) {
-            HStack {
-                VStack(alignment: .leading, spacing: 2) {
-                    Text("Exit Node")
-                        .foregroundColor(Color("TextPeerCard"))
-                    Text(routeViewModel.selectedExitNode?.name ?? "Not Selected")
-                        .font(.footnote)
-                        .foregroundColor(Color("TextSecondary"))
-                }
-                Spacer()
-                Image(systemName: "chevron.right")
-                    .font(.footnote)
-                    .foregroundColor(Color("TextSecondary"))
-            }
-            .padding()
-            .background(Color("BgPeerCard"))
-            .cornerRadius(8)
-        }
-        .padding([.leading, .trailing])
-        .padding(.bottom, 8)
-    }
-}
-
 struct ExitNodeSelectionView: View {
     @ObservedObject var routeViewModel: RoutesViewModel
